requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0
implicit>=0.7.0
scikit-surprise>=1.1.3
torch>=2.0.0
torchvision>=0.15.0

# API & Database
flask>=2.3.0
flask-cors>=4.0.0
sqlalchemy>=2.0.0

# Visualization
plotly>=5.15.0
dash>=2.11.0
dash-bootstrap-components>=1.4.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
